// src/types/generated/graphql.ts
export type Maybe<T> = T | null;
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  timestamptz: any;
  uuid: any;
};

export type Appearance = {
   __typename?: 'appearance';
  created_at: Scalars['timestamptz'];
  id: Scalars['Int'];
  player: Player;
  player_id: Scalars['Int'];
  team: Team;
  team_id: Scalars['Int'];
  updated_at: Scalars['timestamptz'];
};

export type Appearance_Aggregate = {
   __typename?: 'appearance_aggregate';
  aggregate?: Maybe<Appearance_Aggregate_Fields>;
  nodes: Array<Appearance>;
};
// ...
export type Game = {
   __typename?: 'game';
  created_at: Scalars['timestamptz'];
  id: Scalars['Int'];
  matches: Array<Match>;
  matches_aggregate: Match_Aggregate;
  name: Scalars['String'];
  project: Project;
  project_id: Scalars['Int'];
  updated_at: Scalars['timestamptz'];
};
// ...
export type Match = {
   __typename?: 'match';
  created_at: Scalars['timestamptz'];
  date: Scalars['timestamptz'];
  game: Game;
  game_id: Scalars['Int'];
  ratings: Array<Rating>;
  ratings_aggregate: Rating_Aggregate;
  teams: Array<Team>;
  teams_aggregate: Team_Aggregate;
  updated_at: Scalars['timestamptz'];
  uuid: Scalars['uuid'];
};
// ...
export type Match_Aggregate = {
   __typename?: 'match_aggregate';
  aggregate?: Maybe<Match_Aggregate_Fields>;
  nodes: Array<Match>;
};
// ...
export type Match_Insert_Input = {
  created_at?: Maybe<Scalars['timestamptz']>;
  date?: Maybe<Scalars['timestamptz']>;
  game?: Maybe<Game_Obj_Rel_Insert_Input>;
  game_id?: Maybe<Scalars['Int']>;
  ratings?: Maybe<Rating_Arr_Rel_Insert_Input>;
  teams?: Maybe<Team_Arr_Rel_Insert_Input>;
  updated_at?: Maybe<Scalars['timestamptz']>;
  uuid?: Maybe<Scalars['uuid']>;
};
// ...
export type Match_Obj_Rel_Insert_Input = {
  data: Match_Insert_Input;
  on_conflict?: Maybe<Match_On_Conflict>;
};

export type Match_On_Conflict = {
  constraint: Match_Constraint;
  update_columns: Array<Match_Update_Column>;
  where?: Maybe<Match_Bool_Exp>;
};
// ...
export type Player = {
   __typename?: 'player';
  appearances: Array<Appearance>;
  appearances_aggregate: Appearance_Aggregate;
  created_at: Scalars['timestamptz'];
  id: Scalars['Int'];
  name: Scalars['String'];
  project: Project;
  project_id: Scalars['Int'];
  ratings: Array<Rating>;
  ratings_aggregate: Rating_Aggregate;
  updated_at: Scalars['timestamptz'];
};
// ...
export type Player_Insert_Input = {
  appearances?: Maybe<Appearance_Arr_Rel_Insert_Input>;
  created_at?: Maybe<Scalars['timestamptz']>;
  id?: Maybe<Scalars['Int']>;
  name?: Maybe<Scalars['String']>;
  project?: Maybe<Project_Obj_Rel_Insert_Input>;
  project_id?: Maybe<Scalars['Int']>;
  ratings?: Maybe<Rating_Arr_Rel_Insert_Input>;
  updated_at?: Maybe<Scalars['timestamptz']>;
};
// ...
export type Player_Obj_Rel_Insert_Input = {
  data: Player_Insert_Input;
  on_conflict?: Maybe<Player_On_Conflict>;
};

export type Player_On_Conflict = {
  constraint: Player_Constraint;
  update_columns: Array<Player_Update_Column>;
  where?: Maybe<Player_Bool_Exp>;
};
// ...
export type Project = {
   __typename?: 'project';
  created_at: Scalars['timestamptz'];
  games: Array<Game>;
  games_aggregate: Game_Aggregate;
  id: Scalars['Int'];
  name: Scalars['String'];
  players: Array<Player>;
  players_aggregate: Player_Aggregate;
  updated_at: Scalars['timestamptz'];
};
// ...
export type Rating = {
   __typename?: 'rating';
  created_at: Scalars['timestamptz'];
  id: Scalars['Int'];
  match: Match;
  match_uuid: Scalars['uuid'];
  player: Player;
  player_id: Scalars['Int'];
  rating: Scalars['Int'];
  rating_before: Scalars['Int'];
  updated_at: Scalars['timestamptz'];
};

export type Rating_Aggregate = {
   __typename?: 'rating_aggregate';
  aggregate?: Maybe<Rating_Aggregate_Fields>;
  nodes: Array<Rating>;
};

export type Rating_Aggregate_Fields = {
   __typename?: 'rating_aggregate_fields';
  avg?: Maybe<Rating_Avg_Fields>;
  count?: Maybe<Scalars['Int']>;
  max?: Maybe<Rating_Max_Fields>;
  min?: Maybe<Rating_Min_Fields>;
  stddev?: Maybe<Rating_Stddev_Fields>;
  stddev_pop?: Maybe<Rating_Stddev_Pop_Fields>;
  stddev_samp?: Maybe<Rating_Stddev_Samp_Fields>;
  sum?: Maybe<Rating_Sum_Fields>;
  var_pop?: Maybe<Rating_Var_Pop_Fields>;
  var_samp?: Maybe<Rating_Var_Samp_Fields>;
  variance?: Maybe<Rating_Variance_Fields>;
};
// ...
export type Rating_Insert_Input = {
  created_at?: Maybe<Scalars['timestamptz']>;
  id?: Maybe<Scalars['Int']>;
  match?: Maybe<Match_Obj_Rel_Insert_Input>;
  match_uuid?: Maybe<Scalars['uuid']>;
  player?: Maybe<Player_Obj_Rel_Insert_Input>;
  player_id?: Maybe<Scalars['Int']>;
  rating?: Maybe<Scalars['Int']>;
  rating_before?: Maybe<Scalars['Int']>;
  updated_at?: Maybe<Scalars['timestamptz']>;
};
// ...
export enum Result_Type_Enum {
  Draw = 'DRAW',
  Loss = 'LOSS',
  Win = 'WIN'
}
// ...
export type Team = {
   __typename?: 'team';
  appearances: Array<Appearance>;
  appearances_aggregate: Appearance_Aggregate;
  created_at: Scalars['timestamptz'];
  id: Scalars['Int'];
  match: Match;
  match_uuid: Scalars['uuid'];
  result: Result_Type_Enum;
  score: Scalars['Int'];
  updated_at: Scalars['timestamptz'];
};
// ...
export type Team_Aggregate = {
   __typename?: 'team_aggregate';
  aggregate?: Maybe<Team_Aggregate_Fields>;
  nodes: Array<Team>;
};

export type Team_Aggregate_Fields = {
   __typename?: 'team_aggregate_fields';
  avg?: Maybe<Team_Avg_Fields>;
  count?: Maybe<Scalars['Int']>;
  max?: Maybe<Team_Max_Fields>;
  min?: Maybe<Team_Min_Fields>;
  stddev?: Maybe<Team_Stddev_Fields>;
  stddev_pop?: Maybe<Team_Stddev_Pop_Fields>;
  stddev_samp?: Maybe<Team_Stddev_Samp_Fields>;
  sum?: Maybe<Team_Sum_Fields>;
  var_pop?: Maybe<Team_Var_Pop_Fields>;
  var_samp?: Maybe<Team_Var_Samp_Fields>;
  variance?: Maybe<Team_Variance_Fields>;
};
// ...
export type MatchesQueryVariables = {};


export type MatchesQuery = (
  { __typename?: 'query_root' }
  & { match: Array<(
    { __typename?: 'match' }
    & Pick<Match, 'uuid' | 'date' | 'game_id'>
    & { teams: Array<(
      { __typename?: 'team' }
      & Pick<Team, 'id' | 'result' | 'score'>
      & { appearances: Array<(
        { __typename?: 'appearance' }
        & Pick<Appearance, 'player_id'>
      )> }
    )> }
  )> }
);